Abstracts away module into file, organises dirs
