+++ src/components/ThemeToggle.tsx
@@ -3,7 +3,7 @@ import { createTheme, ThemeProvider } from "@mui/material/styles";
 import getDesignTokens from "@/constants/customTheme";
 import CssBaseline from "@mui/material/CssBaseline";
 import useMediaQuery from "@mui/material/useMediaQuery";
-import { Mode } from "@/utils/types";
+import { Mode } from "@/utils/types/mode";
 
 export const ColorModeContext = createContext({ toggleColorMode: () => {} });
