+++ src/utils/theme.d.ts
@@ -1,0 +1,14 @@
+export declare module "@mui/material/styles" {
+  interface Theme {
+    custom: {
+      boxShadow: string;
+      secondaryColor: string;
+    };
+  }
+  interface ThemeOptions {
+    custom?: {
+      boxShadow?: string;
+      secondaryColor?: string;
+    };
+  }
+}
