+++ src/utils/types/country.ts
@@ -102,5 +102,3 @@ export interface PostalCode {
 }
 
 export type Countries = Country[];
-
-export type Mode = "light" | "dark";
