+++ src/utils/types/mode.ts
@@ -1,0 +1,1 @@
+export type Mode = "light" | "dark";
